refactor(go.d/prometheus): move Config and RelabelBlock to config.go
Pure move out of collector.go as the config surface grows; collector.go keeps the CollectorV2 interface methods (including Configuration()) and the constructor. No behavior change.
